refactor: extract `submitCallback` listener
